CLOSE TO FINAL COMMIT
Added code to setup data for Azure cognitive services - Image Recognition.
We added our data processing pipeline upstream of the Azure service and
managed a REST API service into our web app.

Our repo now consists of notebooks that simulate the data,
setup the data, transform the data, and training our models.

The rest of our service is on a react app that will be ported over here.
That app is hosted on AWS, Heroku and domain.com.,